Added events, better dialogue
Added an event for changing the second characters sprite. Made it so that the dialogue text appears letter by letter instead of all at once. The text speed of it appearing can also be changed for each dialogue node. I also added a text font changer to the dialogue node.