Audio manager. Dialogue
Added dialogue manager and click effect to main menu. Added mixer to the audio manager. Moved some stuff with the dialogue such as where the font is to the character name. Also made it so that the default color of all character dialogues is black but can be changed. Also allowed for rich text for the dialogue lines.